Move gallery section to separate component
