+++ src/components/studentEngagementOpportunities.js
@@ -1,119 +1,20 @@
-import React, { useRef } from "react"
+import React from "react"
 import { useWindowSize } from "./hooks"
 import "./studentEngagementOpportunities.scss"
+import ImageGallery from "./imageGallery"
 
 export default function StudentEngagementOpportunities() {
   const { width } = useWindowSize()
-  const img1 = useRef()
-  const img2 = useRef()
-  const img3 = useRef()
-  const gallery = [img1, img2, img3]
-  let curImg = img1
-
-  const handleNextImage = isNext => {
-    const newIndex = isNext ? nextImageIndex() : previousImageIndex()
-    handleStyleUpdates(newIndex)
-    curImg = gallery[newIndex]
-  }
-
-  const previousImageIndex = () => {
-    const curIndex = gallery.findIndex(e => e === curImg)
-    let nextIndex
-
-    if (curIndex === 0) {
-      nextIndex = gallery.length - 1
-    } else {
-      nextIndex = curIndex - 1
-    }
-
-    return nextIndex
-  }
-
-  const nextImageIndex = () => {
-    const curIndex = gallery.findIndex(e => e === curImg)
-    let nextIndex
-
-    if (curIndex === gallery.length - 1) {
-      nextIndex = 0
-    } else {
-      nextIndex = curIndex + 1
-    }
-    console.log(nextIndex)
-    return nextIndex
-  }
-
-  const handleStyleUpdates = newIndex => {
-    gallery.forEach((el, i) => {
-      if (i === newIndex) {
-        el.current.firstChild.classList.remove(
-          "gallery-image__caption--no-display"
-        )
-        el.current.style.zIndex = 3
-      } else {
-        el.current.firstChild.classList.add(
-          "gallery-image__caption--no-display"
-        )
-        el.current.style.zIndex = 2
-      }
-    })
-  }
-
-  const studentEngagementLeft = () => {
-    return (
-      <div className="student-engagement__left">
-          { displayGallery() }
-      </div>
-    )
-  }
-
-  const displayGallery = () => {
-    return (
-      <div className="gallery-wrapper">
-        <div className="gallery">
-          <div className="gallery-container">
-            <div className="gallery-image-wrapper">
-              <div className="gallery-image gallery-image__1" ref={img1}>
-                <div className="gallery-image__caption">
-                  <div className="gallery-image__title">
-                    Film Independent screening + Q&A
-                  </div>
-                </div>
-              </div>
-              <div className="gallery-image gallery-image__2" ref={img2}>
-                <div className="gallery-image__caption gallery-image__caption--no-display">
-                  <div className="gallery-image__title">
-                    PST:LA/LA - A Celebration of Student Culture and Identity at
-                    the Getty Museum
-                  </div>
-                </div>
-              </div>
-              <div className="gallery-image gallery-image__3" ref={img3}>
-                <div className="gallery-image__caption gallery-image__caption--no-display">
-                  <div className="gallery-image__title">
-                    Students present projects to staff on the Paramount lot
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className="gallery-controls">
-              <button onClick={() => handleNextImage(false)}>
-                <div className="gallery-controls__left"></div>
-              </button>
-              <button onClick={() => handleNextImage(true)}>
-                <div className="gallery-controls__right"></div>
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    )
-  }
 
   const displayTextContainer = () => {
     return (
       <div className="text-container">
         <h3>Student Engagement Opportunities</h3>
-        {width < 890 && displayGallery()}
+        {width < 890 && (
+          <div className="student-engagement__left">
+            <ImageGallery />
+          </div>
+        )}
         <p>
           Aside from classroom curriculum, ArtsMatter collaborates with media
           partners, arts organizations, local museums, and content producers to
@@ -131,8 +32,8 @@ export default function StudentEngagementOpportunities() {
         </p>
         <p>
           ​If you're interested in providing educational opportunities for
-          students and schools, please reach out on our <a href="#">Contact</a>{" "}
-          page.
+          students and schools, please reach out on our{" "}
+          <a href="/contact">Contact</a> page.
         </p>
       </div>
     )
@@ -140,7 +41,11 @@ export default function StudentEngagementOpportunities() {
 
   return (
     <div className="student-engagement">
-      {width > 889 && studentEngagementLeft()}
+      {width > 889 && (
+        <div className="student-engagement__left">
+          <ImageGallery />
+        </div>
+      )}
       <div className="student-engagement__right">{displayTextContainer()}</div>
     </div>
   )
